Refactor Object Functor to make it 11x faster
